requirements.txt
Flask
flask-sqlalchemy
requests
python-dotenv
waitress
flask-CORS
scikit-learn
pandas